RF: Move Truncater to a separate module
The logic will get more complicated when we add left and center
styles, so move the truncater logic away to avoid polluting field.py
with helpers.
